refactor CoutryDetail code
